version rendue

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: nfauconn <nfauconn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/07 13:55:46 by nfauconn          #+#    #+#             */
-/*   Updated: 2022/07/10 12:10:33 by nfauconn         ###   ########.fr       */
+/*   Updated: 2022/08/07 01:09:00 by nfauconn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	if (!n || !s1 || !s2)
 		return (0);
 	i = 0;
-	while (s1[i] && s2[i] && (s1[i] == s2[i]) && i < (int)n - 1)
+	while (s1[i] && s2[i] && (s1[i] == s2[i]) && i < n - 1)
 		i++;
 	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
 }
